src/requirements.txt
torch
trl
pandas
numpy
sacrebleu
transformers
tqdm
azure-storage-blob